requirements.txt
Django==3.0.4
djangorestframework==3.11.0
django-htmlmin==0.11.0
